+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Changed
+- Set bounds on options for HoverPreview and SeekPreview
+
 ## [2.4.0] - 2019-09-25
 
 ## Added